tstDisasm-2: size_t warning.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@18458 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Disassembler/testcase/tstDisasm-2.cpp
@@ -560,7 +560,7 @@ static DECLCALLBACK(int) MyDisasInstrRead(RTUINTPTR uSrcAddr, uint8_t *pbDst, ui
                 {
                     memcpy(pbDst, pState->pbNext, pState->cbLeft);
                     pbDst += pState->cbLeft;
-                    cbRead -= pState->cbLeft;
+                    cbRead -= (uint32_t)pState->cbLeft;
                     pState->pbNext += pState->cbLeft;
                     pState->uNextAddr += pState->cbLeft;
                     pState->cbLeft = 0;
